move static data to init.pp & Add datatypes
In our params.pp we define a bunch of variables. On a long point of view
we want to migrate to puppet-strings documentation and data in modules.
puppet-strings can get default values for variables, but only if they
are directly documented in the class. This PR will help us to realize
this.